@@ -21,10 +21,6 @@ exports.initOptionsPage = initOptionsPage;
 exports.loadSettings = loadSettings;
 exports.populateFormFields = populateFormFields;
 exports.setupEventListeners = setupEventListeners;
-exports.setupSearchFunctionality = setupSearchFunctionality;
-exports.performSearch = performSearch;
-exports.highlightMatchingText = highlightMatchingText;
-exports.showNoResultsMessage = showNoResultsMessage;
 exports.setupValidation = setupValidation;
 exports.validatePort = validatePort;
 exports.validateFolder = validateFolder;
@@ -94,7 +90,6 @@ function initOptionsPage() {
     });
     loadSettings();
     setupEventListeners();
-    setupSearchFunctionality();
     setupValidation();
     setupInfoMessages();
     setupTabNavigation();
@@ -171,6 +166,15 @@ function populateFormFields(config) {
     }
     // Trigger validation after populating
     validateAllFields();
+    // Update info messages to reflect current selections
+    const logLevelSelect = document.getElementById("settings-log-level");
+    const formatSelect = document.getElementById("settings-ytdlp-format");
+    if (logLevelSelect) {
+        updateLogLevelInfo(logLevelSelect);
+    }
+    if (formatSelect) {
+        updateFormatInfo(formatSelect);
+    }
 }
 /**
  * Sets up event listeners for form submission and button clicks.
@@ -190,13 +194,8 @@ function setupEventListeners() {
     }
     // Theme toggle functionality
     const themeToggle = document.getElementById("theme-toggle");
-    console.log("Theme toggle element found:", themeToggle);
     if (themeToggle) {
         themeToggle.addEventListener("click", handleThemeToggle);
-        console.log("Theme toggle event listener added");
-    }
-    else {
-        console.error("Theme toggle element not found!");
     }
     const clearHistoryButton = document.getElementById("settings-clear-history");
     if (clearHistoryButton) {
@@ -222,204 +221,6 @@ function setupEventListeners() {
             });
         });
     }
-}
-/**
- * Sets up enhanced search functionality for filtering settings sections.
- */
-function setupSearchFunctionality() {
-    const searchInput = document.getElementById("settings-search");
-    if (!searchInput)
-        return;
-    // Debounced search to improve performance
-    let searchTimeout;
-    searchInput.addEventListener("input", () => {
-        clearTimeout(searchTimeout);
-        searchTimeout = window.setTimeout(() => {
-            performSearch(searchInput.value.trim());
-        }, 300);
-    });
-    // Clear search on escape key
-    searchInput.addEventListener("keydown", (e) => {
-        if (e.key === "Escape") {
-            searchInput.value = "";
-            performSearch("");
-        }
-    });
-    // Add search suggestions
-    setupSearchSuggestions(searchInput);
-}
-/**
- * Performs the actual search and filtering of settings sections.
- */
-function performSearch(query) {
-    const sections = document.querySelectorAll(".settings-group");
-    const searchInput = document.getElementById("settings-search");
-    if (!query) {
-        // Show all sections when search is empty
-        sections.forEach((section) => {
-            section.classList.remove("hidden", "highlighted");
-        });
-        return;
-    }
-    const searchTerms = query
-        .toLowerCase()
-        .split(" ")
-        .filter((term) => term.length > 0);
-    let hasVisibleSections = false;
-    sections.forEach((section) => {
-        var _a, _b, _c, _d;
-        const sectionTitle = ((_b = (_a = section.querySelector(".section-title")) === null || _a === void 0 ? void 0 : _a.textContent) === null || _b === void 0 ? void 0 : _b.toLowerCase()) || "";
-        const sectionContent = ((_c = section.textContent) === null || _c === void 0 ? void 0 : _c.toLowerCase()) || "";
-        const sectionCategory = ((_d = section.getAttribute("data-category")) === null || _d === void 0 ? void 0 : _d.toLowerCase()) || "";
-        // Check if any search term matches
-        const matches = searchTerms.some((term) => sectionTitle.includes(term) ||
-            sectionContent.includes(term) ||
-            sectionCategory.includes(term));
-        if (matches) {
-            section.classList.remove("hidden");
-            section.classList.add("highlighted");
-            hasVisibleSections = true;
-            // Highlight matching text in section titles
-            highlightMatchingText(section, searchTerms);
-        }
-        else {
-            section.classList.add("hidden");
-            section.classList.remove("highlighted");
-        }
-    });
-    // Show "no results" message if no sections match
-    showNoResultsMessage(!hasVisibleSections);
-}
-/**
- * Highlights matching text in section titles.
- */
-function highlightMatchingText(section, searchTerms) {
-    const titleElement = section.querySelector(".section-title");
-    if (!titleElement)
-        return;
-    let titleText = titleElement.textContent || "";
-    const originalText = titleText;
-    // Remove existing highlights
-    titleText = titleText.replace(/<mark class="search-highlight">(.*?)<\/mark>/g, "$1");
-    // Add new highlights
-    searchTerms.forEach((term) => {
-        const regex = new RegExp("(" + term + ")", "gi");
-        titleText = titleText.replace(regex, '<mark class="search-highlight">$1</mark>');
-    });
-    if (titleText !== originalText) {
-        titleElement.innerHTML = titleText;
-    }
-}
-/**
- * Shows or hides the "no results" message.
- */
-function showNoResultsMessage(show) {
-    let noResultsElement = document.getElementById("no-results-message");
-    if (show) {
-        if (!noResultsElement) {
-            noResultsElement = document.createElement("div");
-            noResultsElement.id = "no-results-message";
-            noResultsElement.className = "no-results-message";
-            noResultsElement.innerHTML =
-                '<div class="no-results-content">' +
-                    '<span class="no-results-icon">Search</span>' +
-                    "<p>No settings found matching your search.</p>" +
-                    "<p>Try different keywords or check the spelling.</p>" +
-                    "</div>";
-            const container = document.querySelector(".settings-container");
-            if (container) {
-                container.appendChild(noResultsElement);
-            }
-        }
-        noResultsElement.style.display = "block";
-    }
-    else if (noResultsElement) {
-        noResultsElement.style.display = "none";
-    }
-}
-/**
- * Sets up search suggestions for common settings.
- */
-function setupSearchSuggestions(searchInput) {
-    var _a;
-    const suggestions = [
-        "port",
-        "server",
-        "download",
-        "folder",
-        "format",
-        "video",
-        "audio",
-        "playlist",
-        "history",
-        "debug",
-        "log",
-        "level",
-        "behavior",
-        "actions",
-        "resume",
-        "clear",
-        "restart",
-        "logs",
-        "errors",
-    ];
-    // Create suggestion dropdown
-    const suggestionDropdown = document.createElement("div");
-    suggestionDropdown.id = "search-suggestions";
-    suggestionDropdown.className = "search-suggestions";
-    (_a = searchInput.parentNode) === null || _a === void 0 ? void 0 : _a.appendChild(suggestionDropdown);
-    searchInput.addEventListener("focus", () => {
-        if (searchInput.value.trim()) {
-            showSuggestions(searchInput.value.trim(), suggestions, suggestionDropdown);
-        }
-    });
-    searchInput.addEventListener("input", () => {
-        const query = searchInput.value.trim();
-        if (query.length > 0) {
-            showSuggestions(query, suggestions, suggestionDropdown);
-        }
-        else {
-            suggestionDropdown.style.display = "none";
-        }
-    });
-    // Hide suggestions when clicking outside
-    document.addEventListener("click", (e) => {
-        if (!searchInput.contains(e.target) &&
-            !suggestionDropdown.contains(e.target)) {
-            suggestionDropdown.style.display = "none";
-        }
-    });
-}
-/**
- * Shows search suggestions based on the current query.
- */
-function showSuggestions(query, suggestions, dropdown) {
-    const matchingSuggestions = suggestions
-        .filter((suggestion) => suggestion.toLowerCase().includes(query.toLowerCase()))
-        .slice(0, 5); // Limit to 5 suggestions
-    if (matchingSuggestions.length === 0) {
-        dropdown.style.display = "none";
-        return;
-    }
-    dropdown.innerHTML = matchingSuggestions
-        .map((suggestion) => '<div class="suggestion-item" data-suggestion="' +
-        suggestion +
-        '">' +
-        suggestion +
-        "</div>")
-        .join("");
-    dropdown.style.display = "block";
-    // Add click handlers for suggestions
-    dropdown.querySelectorAll(".suggestion-item").forEach((item) => {
-        item.addEventListener("click", () => {
-            const searchInput = document.getElementById("settings-search");
-            if (searchInput) {
-                searchInput.value = item.getAttribute("data-suggestion") || "";
-                performSearch(searchInput.value);
-                dropdown.style.display = "none";
-            }
-        });
-    });
 }
 /**
  * Sets up comprehensive validation for all form fields with helpful messages.
@@ -735,9 +536,9 @@ function updateLogLevelInfo(select) {
     if (!infoText)
         return;
     const levelInfo = {
-        ERROR: "Only error messages will be logged",
-        INFO: "Normal level provides essential information",
-        DEBUG: "Verbose logging for troubleshooting",
+        error: "Only error messages will be logged",
+        info: "Normal level provides essential information",
+        debug: "Verbose logging for troubleshooting",
     };
     infoText.textContent = levelInfo[select.value] || "Select a log level";
 }
@@ -1014,17 +815,17 @@ function loadErrorHistory() {
 function applyOptionsTheme(forceTheme) {
     const isDark = forceTheme === "dark" ||
         (!forceTheme && window.matchMedia("(prefers-color-scheme: dark)").matches);
-    console.log("Applying theme:", { forceTheme, isDark });
-    document.documentElement.classList.toggle("dark-theme", isDark);
+    // Applying theme
+    document.body.classList.toggle("dark-theme", isDark);
     // Update header icon based on theme
     const headerIcon = document.getElementById("options-header-icon");
     if (headerIcon) {
         const currentSrc = headerIcon.src;
         const isCurrentlyDark = currentSrc.includes("darkicon");
-        console.log("Header icon update:", { currentSrc, isCurrentlyDark, isDark });
+        // Header icon update
         if (isDark !== isCurrentlyDark) {
             const newSrc = currentSrc.replace(isCurrentlyDark ? "darkicon48.png" : "icon48.png", isDark ? "darkicon48.png" : "icon48.png");
-            console.log("Updating icon from", currentSrc, "to", newSrc);
+            // Updating icon
             headerIcon.src = newSrc;
         }
     }
@@ -1035,12 +836,12 @@ function applyOptionsTheme(forceTheme) {
  */
 function handleThemeToggle() {
     return __awaiter(this, void 0, void 0, function* () {
-        console.log("Theme toggle clicked!");
+        // Theme toggle clicked
         try {
             // Get current theme from storage
             const result = yield chrome.storage.local.get("theme");
             const currentTheme = result.theme;
-            console.log("Current theme from storage:", currentTheme);
+            // Current theme from storage
             // Determine new theme
             let newTheme;
             if (currentTheme === "dark") {
@@ -1054,13 +855,13 @@ function handleThemeToggle() {
                 const systemPrefersDark = window.matchMedia("(prefers-color-scheme: dark)").matches;
                 newTheme = systemPrefersDark ? "light" : "dark";
             }
-            console.log("New theme will be:", newTheme);
+            // New theme will be
             // Save new theme to storage
             yield chrome.storage.local.set({ theme: newTheme });
             // Apply the new theme
             applyOptionsTheme(newTheme);
             // Log the theme change
-            console.log(`Theme changed to: ${newTheme}`);
+            // Theme changed
         }
         catch (error) {
             console.error("Error toggling theme:", error);
