@@ -286,10 +286,9 @@ describe("Background Logic: discoverServerPort", () => {
         expect(mockCheckStatus).toHaveBeenCalledWith(5001);
         expect(result).toBe(5001);
     }));
-    // Enhanced timeout testing
+    // Optimized timeout testing - using immediate rejection instead of setTimeout
     it("should handle timeout for cached port check", () => __awaiter(void 0, void 0, void 0, function* () {
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)) // Longer than timeout
-        );
+        mockCheckStatus.mockImplementation(() => new Promise((_, reject) => reject(new Error("Timeout"))));
         const storageService = createMockStorageService({
             getPort: jest.fn().mockResolvedValue(5013),
             setPort: jest.fn().mockResolvedValue(undefined),
@@ -300,8 +299,7 @@ describe("Background Logic: discoverServerPort", () => {
         expect(storageService.setPort).toHaveBeenCalledWith(null);
     }));
     it("should handle timeout during port scanning", () => __awaiter(void 0, void 0, void 0, function* () {
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)) // Longer than timeout
-        );
+        mockCheckStatus.mockImplementation(() => new Promise((_, reject) => reject(new Error("Timeout"))));
         const storageService = createMockStorageService({
             getPort: jest.fn().mockResolvedValue(null),
         });
@@ -354,69 +352,33 @@ describe("Background Logic: discoverServerPort", () => {
         const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5003, false, 1000);
         expect(result).toBeNull();
     }));
-    it("should handle timeout with different error messages", () => __awaiter(void 0, void 0, void 0, function* () {
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(5013),
-            setPort: jest.fn().mockResolvedValue(undefined),
-        });
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5010, false, 1000);
-        expect(result).toBeNull();
-    }));
-    it("should handle timeout during scanning with different error messages", () => __awaiter(void 0, void 0, void 0, function* () {
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 1000);
-        expect(result).toBeNull();
-    }));
-    it("should handle storage service without setPort method during cached port failure", () => __awaiter(void 0, void 0, void 0, function* () {
-        mockCheckStatus.mockResolvedValue(false);
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(5013),
-            setPort: undefined, // Remove setPort method
-        });
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5010, false, 1000);
-        expect(result).toBeNull();
-    }));
-    it("should handle storage service without setPort method during timeout", () => __awaiter(void 0, void 0, void 0, function* () {
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(5013),
-            setPort: undefined, // Remove setPort method
-        });
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5010, false, 1000);
-        expect(result).toBeNull();
-    }));
-    // Batch processing edge case tests
+    // Batch processing tests
     it("should handle batch processing edge case with exact batch size", () => __awaiter(void 0, void 0, void 0, function* () {
         mockCheckStatus.mockResolvedValue(false);
         const storageService = createMockStorageService({
             getPort: jest.fn().mockResolvedValue(null),
         });
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, // 5 ports, batch size is 5
+        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, // 5 ports, batch size is 5
         false, 1000);
-        expect(result).toBeNull();
+        expect(mockCheckStatus).toHaveBeenCalledTimes(5);
     }));
     it("should handle batch processing with remainder ports", () => __awaiter(void 0, void 0, void 0, function* () {
         mockCheckStatus.mockResolvedValue(false);
         const storageService = createMockStorageService({
             getPort: jest.fn().mockResolvedValue(null),
         });
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5007, // 7 ports, batch size is 5 (remainder of 2)
+        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5007, // 7 ports, batch size is 5, remainder is 2
         false, 1000);
-        expect(result).toBeNull();
+        expect(mockCheckStatus).toHaveBeenCalledTimes(7);
     }));
-    it("should handle progress tracking with empty range", () => __awaiter(void 0, void 0, void 0, function* () {
-        mockCheckStatus.mockResolvedValue(false);
-        const onProgress = jest.fn();
+    it("should scan ports in batches efficiently", () => __awaiter(void 0, void 0, void 0, function* () {
         const storageService = createMockStorageService({
             getPort: jest.fn().mockResolvedValue(null),
         });
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5001, // Single port range
-        false, 1000, onProgress);
-        expect(onProgress).toHaveBeenCalled();
+        mockCheckStatus.mockResolvedValue(false);
+        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5010, // 10 ports, 2 batches of 5
+        false, 1000);
+        expect(mockCheckStatus).toHaveBeenCalledTimes(10);
     }));
     // Null/undefined handling tests
     it("should handle null port discovery in batch processing", () => __awaiter(void 0, void 0, void 0, function* () {
@@ -451,7 +413,7 @@ describe("Background Logic: discoverServerPort", () => {
             getPort: jest.fn().mockResolvedValue(null),
         });
         mockCheckStatus.mockResolvedValue(false);
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 2000);
+        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 1000);
         expect(result).toBeNull();
     }));
     it("should handle storage service without setPort method during scan", () => __awaiter(void 0, void 0, void 0, function* () {
@@ -460,7 +422,7 @@ describe("Background Logic: discoverServerPort", () => {
             setPort: undefined, // Remove setPort method
         });
         mockCheckStatus.mockResolvedValue(true);
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 2000);
+        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 1000);
         expect(result).toBe(5001);
     }));
     it("should handle errors during port scanning gracefully", () => __awaiter(void 0, void 0, void 0, function* () {
@@ -468,73 +430,19 @@ describe("Background Logic: discoverServerPort", () => {
             getPort: jest.fn().mockResolvedValue(null),
         });
         mockCheckStatus.mockRejectedValue(new Error("Network error"));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 2000);
-        expect(result).toBeNull();
-    }));
-    it("should scan ports in batches efficiently", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5010, false, 2000);
-        // Should check all ports in the range
-        expect(mockCheckStatus).toHaveBeenCalledTimes(10);
-    }));
-    // Boundary condition tests
-    it("should handle timeout with different error messages", () => __awaiter(void 0, void 0, void 0, function* () {
-        const cachedPort = 5013;
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(cachedPort),
-        });
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5020, false, 1000);
-        expect(result).toBeNull();
-    }), 10000); // Increase test timeout
-    it("should handle timeout during scanning with different error messages", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
         const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 1000);
         expect(result).toBeNull();
-    }), 10000); // Increase test timeout
-    it("should handle storage service without setPort method during cached port failure", () => __awaiter(void 0, void 0, void 0, function* () {
-        const cachedPort = 5013;
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(cachedPort),
-            setPort: undefined, // Remove setPort method
-        });
-        mockCheckStatus.mockRejectedValue(new Error("Connection failed"));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5020, false, 2000);
-        expect(result).toBeNull();
     }));
-    it("should handle storage service without setPort method during timeout", () => __awaiter(void 0, void 0, void 0, function* () {
-        const cachedPort = 5013;
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(cachedPort),
-            setPort: undefined, // Remove setPort method
-        });
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5020, false, 1000);
-        expect(result).toBeNull();
-    }), 10000); // Increase test timeout
-    it("should handle batch processing edge case with exact batch size", () => __awaiter(void 0, void 0, void 0, function* () {
+    // Progress tracking tests
+    it("should handle progress tracking with empty range", () => __awaiter(void 0, void 0, void 0, function* () {
+        mockCheckStatus.mockResolvedValue(false);
+        const onProgress = jest.fn();
         const storageService = createMockStorageService({
             getPort: jest.fn().mockResolvedValue(null),
         });
-        mockCheckStatus.mockResolvedValue(false);
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, // 5 ports, batch size is 5
-        false, 2000);
-        expect(mockCheckStatus).toHaveBeenCalledTimes(5);
-    }));
-    it("should handle batch processing with remainder ports", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5007, // 7 ports, batch size is 5, remainder is 2
-        false, 2000);
-        expect(mockCheckStatus).toHaveBeenCalledTimes(7);
+        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5001, // Single port range
+        false, 1000, onProgress);
+        expect(onProgress).toHaveBeenCalled();
     }));
     it("should handle progress tracking with negative batch length", () => __awaiter(void 0, void 0, void 0, function* () {
         const storageService = createMockStorageService({
@@ -543,255 +451,7 @@ describe("Background Logic: discoverServerPort", () => {
         mockCheckStatus.mockResolvedValue(false);
         const onProgress = jest.fn();
         const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5000, // Invalid range (negative)
-        false, 2000, onProgress);
+        false, 1000, onProgress);
         expect(result).toBeNull();
-    }));
-    it("should handle null port discovery in batch processing", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(null);
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 2000);
-        expect(result).toBeNull();
-    }));
-    it("should handle undefined port discovery in batch processing", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(undefined);
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 2000);
-        expect(result).toBeNull();
-    }));
-    it("should handle mixed null/undefined results in batch processing", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus
-            .mockResolvedValueOnce(null)
-            .mockResolvedValueOnce(undefined)
-            .mockResolvedValueOnce(5003);
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 2000);
-        expect(result).toBe(5003);
-    }));
-    // Additional tests to address remaining survived mutants
-    it("should validate specific error message content for timeout errors", () => __awaiter(void 0, void 0, void 0, function* () {
-        const cachedPort = 5013;
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(cachedPort),
-            setPort: jest.fn().mockResolvedValue(undefined),
-        });
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5010, false, 1000 // Short timeout
-        );
-        expect(result).toBeNull();
-        expect(storageService.setPort).toHaveBeenCalledWith(null);
-    }), 10000);
-    it("should validate specific error message content for scanning timeout errors", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 1000 // Short timeout
-        );
-        expect(result).toBeNull();
-    }), 10000);
-    it("should handle boundary condition with exact port range length", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, // 5 ports, batch size is 5
-        false, 2000);
-        expect(result).toBeNull();
-    }));
-    it("should handle boundary condition with remainder ports", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5007, // 7 ports, batch size is 5, remainder is 2
-        false, 2000);
-        expect(result).toBeNull();
-    }));
-    it("should handle exception in catch block with proper error handling", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(5013),
-        });
-        mockCheckStatus.mockRejectedValue(new Error("Network error"));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5020, false, 2000);
-        expect(result).toBeNull();
-    }));
-    it("should handle null/undefined foundPort with proper conditional logic", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5003, false, 2000);
-        expect(result).toBeNull();
-    }));
-    it("should handle progress tracking with proper assignment operator", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        const onProgress = jest.fn();
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 2000, onProgress);
-        expect(onProgress).toHaveBeenCalled();
-    }));
-    it("should handle progress tracking with multiple batches", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        const onProgress = jest.fn();
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5010, // 10 ports, 2 batches of 5
-        false, 2000, onProgress);
-        expect(onProgress).toHaveBeenCalled();
-    }));
-    it("should handle progress tracking with single batch", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        const onProgress = jest.fn();
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5003, // 3 ports, single batch
-        false, 2000, onProgress);
-        expect(onProgress).toHaveBeenCalled();
-    }));
-    it("should handle progress tracking with empty range", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        const onProgress = jest.fn();
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5001, // Single port range
-        false, 2000, onProgress);
-        expect(onProgress).toHaveBeenCalled();
-    }));
-    it("should handle progress tracking with negative batch length", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        const onProgress = jest.fn();
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5001, // Single port range
-        false, 2000, onProgress);
-        expect(onProgress).toHaveBeenCalled();
-    }));
-    it("should handle null port discovery in batch processing", () => __awaiter(void 0, void 0, void 0, function* () {
-        mockCheckStatus.mockResolvedValue(null);
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5003, false, 1000);
-        expect(result).toBeNull();
-    }));
-    it("should handle undefined port discovery in batch processing", () => __awaiter(void 0, void 0, void 0, function* () {
-        mockCheckStatus.mockResolvedValue(undefined);
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5003, false, 1000);
-        expect(result).toBeNull();
-    }));
-    it("should handle mixed null/undefined results in batch processing", () => __awaiter(void 0, void 0, void 0, function* () {
-        mockCheckStatus
-            .mockResolvedValueOnce(false)
-            .mockResolvedValueOnce(false)
-            .mockResolvedValueOnce(true);
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5003, false, 1000);
-        expect(result).toBe(5003); // Third port (5003) should be found
-    }));
-    it("should handle port scanning when no port is found", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 2000);
-        expect(result).toBeNull();
-    }));
-    it("should handle storage service without setPort method during scan", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-            setPort: undefined, // Remove setPort method
-        });
-        mockCheckStatus.mockResolvedValue(true);
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 2000);
-        expect(result).toBe(5001);
-    }));
-    it("should handle errors during port scanning gracefully", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockRejectedValue(new Error("Scan error"));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 2000);
-        expect(result).toBeNull();
-    }));
-    it("should handle timeout with different error messages", () => __awaiter(void 0, void 0, void 0, function* () {
-        const cachedPort = 5013;
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(cachedPort),
-        });
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5020, false, 1000);
-        expect(result).toBeNull();
-    }), 10000);
-    it("should handle timeout during scanning with different error messages", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, false, 1000);
-        expect(result).toBeNull();
-    }), 10000);
-    it("should handle storage service without setPort method during cached port failure", () => __awaiter(void 0, void 0, void 0, function* () {
-        const cachedPort = 5013;
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(cachedPort),
-            setPort: undefined, // Remove setPort method
-        });
-        mockCheckStatus.mockRejectedValue(new Error("Connection failed"));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5020, false, 2000);
-        expect(result).toBeNull();
-    }));
-    it("should handle storage service without setPort method during timeout", () => __awaiter(void 0, void 0, void 0, function* () {
-        const cachedPort = 5013;
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(cachedPort),
-            setPort: undefined, // Remove setPort method
-        });
-        mockCheckStatus.mockImplementation(() => new Promise((resolve) => setTimeout(() => resolve(true), 3000)));
-        const result = yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5020, false, 1000);
-        expect(result).toBeNull();
-    }), 10000);
-    it("should handle batch processing edge case with exact batch size", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5005, // 5 ports, batch size is 5
-        false, 2000);
-        expect(mockCheckStatus).toHaveBeenCalledTimes(5);
-    }));
-    it("should handle batch processing with remainder ports", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5007, // 7 ports, batch size is 5, remainder is 2
-        false, 2000);
-        expect(mockCheckStatus).toHaveBeenCalledTimes(7);
-    }));
-    it("should scan ports in batches efficiently", () => __awaiter(void 0, void 0, void 0, function* () {
-        const storageService = createMockStorageService({
-            getPort: jest.fn().mockResolvedValue(null),
-        });
-        mockCheckStatus.mockResolvedValue(false);
-        yield (0, background_logic_1.discoverServerPort)(storageService, mockCheckStatus, 5001, 5010, // 10 ports, 2 batches of 5
-        false, 2000);
-        expect(mockCheckStatus).toHaveBeenCalledTimes(10);
     }));
 });
