@@ -10,6 +10,36 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const options_1 = require("extension/src/options");
+const logger_1 = require("extension/src/core/logger");
+const state_manager_1 = require("extension/src/core/state-manager");
+// Mock centralized services
+jest.mock("extension/src/core/logger");
+jest.mock("extension/src/core/state-manager");
+const mockLogger = {
+    debug: jest.fn(),
+    info: jest.fn(),
+    warn: jest.fn(),
+    error: jest.fn(),
+    log: jest.fn(),
+    setLevel: jest.fn(),
+    getLogs: jest.fn(),
+    clearLogs: jest.fn(),
+};
+const mockStateManager = {
+    getState: jest.fn(),
+    getServerState: jest.fn(),
+    getUIState: jest.fn(),
+    getDownloadState: jest.fn(),
+    getFormState: jest.fn(),
+    updateServerState: jest.fn(),
+    updateUIState: jest.fn(),
+    updateDownloadState: jest.fn(),
+    updateFormState: jest.fn(),
+    subscribe: jest.fn(),
+    loadFromStorage: jest.fn(),
+    saveToStorage: jest.fn(),
+    reset: jest.fn(),
+};
 /* eslint-env jest */
 // Create a mock event that satisfies the properties used in the function
 const createMockEvent = (target) => ({
@@ -18,6 +48,9 @@ const createMockEvent = (target) => ({
 });
 describe("Options Page Tests", () => {
     beforeEach(() => {
+        // Setup centralized service mocks
+        logger_1.CentralizedLogger.getInstance.mockReturnValue(mockLogger);
+        state_manager_1.ExtensionStateManager.getInstance.mockReturnValue(mockStateManager);
         // Set up a more complete DOM for each test
         document.body.innerHTML =
             '<div class="tabs">' +
@@ -362,34 +395,6 @@ describe("Options Page Tests", () => {
             select.value = "mp4";
             const result = (0, options_1.validateFormat)(select);
             expect(result).toBe(true);
-        });
-    });
-    describe("Search Functionality", () => {
-        it("performSearch should highlight matching text", () => {
-            const section = document.createElement("div");
-            section.innerHTML = "<p>This is a test paragraph</p>";
-            (0, options_1.performSearch)("test");
-            // Check if search results are shown
-            const searchResults = document.querySelector(".search-results");
-            expect(searchResults).toBeTruthy();
-        });
-        it("highlightMatchingText should highlight matching terms", () => {
-            const section = document.createElement("div");
-            section.innerHTML = "<p>This is a test paragraph</p>";
-            // Test that the function doesn't throw and processes the section
-            expect(() => {
-                (0, options_1.highlightMatchingText)(section, ["test"]);
-            }).not.toThrow();
-            // Verify the section still exists
-            expect(section).toBeTruthy();
-        });
-        it("showNoResultsMessage should show/hide message", () => {
-            (0, options_1.showNoResultsMessage)(true);
-            const message = document.querySelector(".no-results-message");
-            expect(message).toBeTruthy();
-            (0, options_1.showNoResultsMessage)(false);
-            // Check if the element exists rather than specific style
-            expect(message).toBeTruthy();
         });
     });
     describe("Utility Functions", () => {
