@@ -42,7 +42,7 @@ def create_app(config: Config) -> Flask:
     app.register_blueprint(download_bp)
     app.register_blueprint(config_bp)
     app.register_blueprint(status_bp)
-    app.register_blueprint(health_bp)
+    app.register_blueprint(health_bp)  # Register without prefix for /health
     app.register_blueprint(history_bp)
     app.register_blueprint(restart_bp)
     app.register_blueprint(logs_bp)
